Organized logic, fixed a few bugs
Converted the Sensors to the Observer pattern so that the simulator
triggers the Sensor, not the ChronoTimer (which is what the requirements
seem to want).
Moved Individual Event logic out of the ChronoTimer class and into a
IndEventController class.
Added a stub for a Parallel Individual Event Controller.
Moved the checks for if the ChronoTimer is powered on into the
ChronoTimer to be more accurate.