Test Updates
I've made some of the tests more comprehensive for the different types
of events from the perspective of the Chronotimer, and I also tested and
fixed some parts of the Swap function. I added a couple extra getters
into Chronotimer and Run so I could look at some of the variables while
testing. I've left the ones that still need work with Todo's.